Remove ConsumerObserver
