+++ src/main/java/io/appform/dropwizard/actors/connectivity/RMQConnection.java
@@ -32,7 +32,6 @@ import io.appform.dropwizard.actors.TtlConfig;
 import io.appform.dropwizard.actors.actor.ActorConfig;
 import io.appform.dropwizard.actors.base.utils.NamingUtils;
 import io.appform.dropwizard.actors.config.RMQConfig;
-import io.appform.dropwizard.actors.metrics.ConsumerMetricObserver;
 import io.appform.dropwizard.actors.metrics.PublishMetricObserver;
 import io.appform.dropwizard.actors.observers.RMQPublishObserver;
 import io.appform.dropwizard.actors.observers.TerminalObserver;
@@ -279,7 +278,7 @@ public class RMQConnection implements Managed {
             }
             rootObserver = observer.setNext(rootObserver);
         }
-        rootObserver = new ConsumerMetricObserver(config, metricRegistry).setNext(rootObserver);
+//        rootObserver = new ConsumerMetricObserver(config, metricRegistry).setNext(rootObserver);
         this.rootObserver = new PublishMetricObserver(config, metricRegistry).setNext(rootObserver);
     }
 }
