+++ src/main/java/io/appform/dropwizard/actors/observers/RMQPublishObserver.java
@@ -16,17 +16,26 @@ public abstract class RMQPublishObserver {
         this.next = next;
     }
 
-    public abstract <T> T execute(final PublishObserverContext context, Supplier<T> supplier);
+    public abstract <T> T executePublish(final PublishObserverContext context, Supplier<T> supplier);
+
+    public abstract <T> T executeConsume(final PublishObserverContext context, Supplier<T> supplier);
 
     public final RMQPublishObserver setNext(final RMQPublishObserver next) {
         this.next = next;
         return this;
     }
 
-    protected final <T> T proceed(final PublishObserverContext context, final Supplier<T> supplier) {
+    protected final <T> T proceedPublish(final PublishObserverContext context, final Supplier<T> supplier) {
         if (null == next) {
             return supplier.get();
         }
-        return next.execute(context, supplier);
+        return next.executePublish(context, supplier);
+    }
+
+    protected final <T> T proceedConsume(final PublishObserverContext context, final Supplier<T> supplier) {
+        if (null == next) {
+            return supplier.get();
+        }
+        return next.executeConsume(context, supplier);
     }
 }
