+++ src/test/java/io/appform/dropwizard/actors/metrics/PublishMetricObserverTest.java
@@ -40,7 +40,7 @@ class PublishMetricObserverTest {
         config.setMetricConfig(MetricConfig.builder().enabledForAll(false).build());
         val publishMetricObserver = new PublishMetricObserver(config, metricRegistry);
         assertEquals(terminate(),
-                publishMetricObserver.execute(PublishObserverContext.builder().build(), this::terminate));
+                publishMetricObserver.executePublish(PublishObserverContext.builder().build(), this::terminate));
     }
 
     @Test
@@ -49,7 +49,7 @@ class PublishMetricObserverTest {
                 .operation(PublishOperations.PUBLISH.name())
                 .queueName("default")
                 .build();
-        assertEquals(terminate(), publishMetricObserver.execute(context, this::terminate));
+        assertEquals(terminate(), publishMetricObserver.executePublish(context, this::terminate));
         val key = MetricKeyData.builder().operation(context.getOperation()).queueName(context.getQueueName()).build();
         validateMetrics(publishMetricObserver.getMetricCache().get(key), 1, 0);
     }
@@ -60,7 +60,7 @@ class PublishMetricObserverTest {
                 .operation(PublishOperations.PUBLISH.name())
                 .queueName("default")
                 .build();
-        assertThrows(RuntimeException.class, () -> publishMetricObserver.execute(context, this::terminateWithException));
+        assertThrows(RuntimeException.class, () -> publishMetricObserver.executePublish(context, this::terminateWithException));
         val key = MetricKeyData.builder().operation(context.getOperation()).queueName(context.getQueueName()).build();
         validateMetrics(publishMetricObserver.getMetricCache().get(key), 0, 1);
     }
